Update WorkPlan, raws, UPH, Running

+++ Andon.Biz.Logic/MES_MSG_LINE_PRODUCT.cs
@@ -20,7 +20,7 @@ namespace iAndon.Biz.Logic
         public string PRODUCT_CODE { get; set; }
         public string PRODUCT_NAME { get; set; }
         public decimal PLAN_QUANTITY { get; set; }
-        public short HEAD_COUNT { get; set; }
+        public int HEAD_COUNT { get; set; }
         public decimal TAKT_TIME { get; set; }
         public decimal UPH { get; set; }
         public decimal UPPH { get; set; }
